[ONL-8261] chore: Migrated ec-badge needed in filter-popover
