+++ src/components/ec-badge/index.ts
