+++ src/components/ec-badge/types.ts
@@ -1,0 +1,8 @@
+import { AlertType as BadgeType } from '../ec-alert/types';
+
+export { BadgeType };
+
+export interface BadgeProps {
+  type?: BadgeType,
+  value: string | number,
+}
